Exception Handling

+++ src/main/java/com/stackroute/MovieApp/service/MovieServiceImpl.java
@@ -1,10 +1,13 @@
 package com.stackroute.MovieApp.service;
 
 import com.stackroute.MovieApp.domain.Movie;
+import com.stackroute.MovieApp.exception.MovieAlreadyExistsException;
+import com.stackroute.MovieApp.exception.MovieNotFoundException;
 import com.stackroute.MovieApp.repository.MovieRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,20 +18,22 @@ public class MovieServiceImpl implements MovieService {
 
   @Autowired
   public MovieServiceImpl(MovieRepository movieRepository){
+    super();
     this.movieRepository = movieRepository;
   }
 
   @Override
-  public Movie saveNewMovie(Movie movie)  {
-//    if(movieRepository.existsById(movie.getId())){
-//      throw new Exception("Movie already exits");
-//    }
+  public Movie saveNewMovie(Movie movie) throws MovieAlreadyExistsException {
+
+    if(movieRepository.existsById(movie.getId())){
+      throw new MovieAlreadyExistsException("Movie Already Exists");
+    }
     Movie savedMovie = movieRepository.save(movie);
-//    if(savedMovie == null)
-//    {
-//      throw new Exception("Movie already exists.");
-//    }
-    return savedMovie;
+    if (savedMovie == null){
+      throw new MovieAlreadyExistsException("Movie already exists");
+    }
+
+   return savedMovie;
   }
 
   @Override
@@ -37,31 +42,32 @@ public class MovieServiceImpl implements MovieService {
   }
 
   @Override
-  public Optional<Movie> getById(int id) {
+  public Optional<Movie> getById(int id) throws MovieNotFoundException {
     Optional<Movie> movieId = movieRepository.findById(id);
-//    if (movieId.isEmpty()){
-//      throw new Exception("Movie not found");
-//    }
-    return movieId;
+    if (movieId.isPresent()){
+       return movieId;
+    }else {
+      throw new MovieNotFoundException("Movie Not Found");
+    }
   }
 
   @Override
-  public boolean deleteById(int id){
+  public boolean deleteById(int id) throws MovieNotFoundException{
     Optional<Movie> movieId = movieRepository.findById(id);
-//    if (movieId.isEmpty()){
-//      throw new Exception("Movie not found");
-//    }
+    if (movieId.isEmpty()){
+      throw new MovieNotFoundException("Movie not found");
+    }
      movieRepository.deleteById(id);
      return true;
 
   }
 
   @Override
-  public Movie updateById(Movie movie, int id) {
+  public Movie updateById(Movie movie, int id) throws MovieNotFoundException {
     Optional<Movie> userOptional = movieRepository.findById(id);
-//    if(userOptional.isEmpty()){
-//      throw new Exception("Track not found!");
-//    }
+    if(userOptional.isEmpty()){
+      throw new MovieNotFoundException("Track not found!");
+    }
     movie.setId(id);
     movieRepository.save(movie);
     return movie;
